Move filter into a separated file
